requirements.txt
dj-database-url==0.5.0
Django==1.11
django-heroku==0.3.1
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.6
whitenoise==3.2
